Move App into components folder
